requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs
crawl4ai
playwright
sentence-transformers
numpy
openai>=1.0
prompt-toolkit
questionary
rich
ollama
jedi
pyyaml
